+++ src/api/types/emails.ts
@@ -3,3 +3,19 @@ export interface PostSendEmailParams {
   subject: string;
   content?: string;
 }
+
+export interface putEmailsParams {
+  clientId: number;
+  data: FormData;
+}
+
+export interface putEmailsData {
+  recipients: string[];
+  ccs: string[];
+}
+
+export interface GetEmailsResponse {
+  recipients: string[];
+  ccs: string[];
+  signatureImageUrl: string | null;
+}
