+++ src/api/types/keywords.ts
@@ -1,0 +1,23 @@
+export interface Keyword {
+  keywordId: number;
+  keywordName: string;
+}
+
+export type GetKeywordsResponse = {
+  [key: string]: Keyword[];
+};
+
+interface KeywordsByCategory {
+  SELF: string[];
+  COMPETITOR: string[];
+  INDUSTRY: string[];
+}
+
+export interface putKeywordsData {
+  keywordsByCategory: KeywordsByCategory;
+}
+
+export interface PutKeywordsParams {
+  clientId: number;
+  data: putKeywordsData;
+}
